backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
alembic==1.14.1
psycopg[binary]==3.2.9
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.10.1
httpx==0.28.1
redis==5.2.1
rq==2.1.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==2.18.0
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-sqlalchemy==0.51b0
edge-tts==7.2.7
email-validator==2.2.0
python-multipart==0.0.20
